backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
slowapi
pinecone